Split logger.hpp into verbosity_levels.hpp containing the enumeration + related functions and logging.hpp containing the actual logging function.
