+++ src/main_train.cpp
@@ -12,8 +12,9 @@
 
 #include "plssvm/detail/cmd/data_set_variants.hpp"  // plssvm::detail::cmd::data_set_factory
 #include "plssvm/detail/cmd/parser_train.hpp"       // plssvm::detail::cmd::parser_train
-#include "plssvm/detail/logger.hpp"                 // plssvm::detail::log, plssvm::verbosity_level
+#include "plssvm/detail/logging.hpp"                // plssvm::detail::log
 #include "plssvm/detail/performance_tracker.hpp"    // plssvm::detail::tracking_entry, PLSSVM_DETAIL_PERFORMANCE_TRACKER_SAVE
+#include "plssvm/verbosity_levels.hpp"              // plssvm::verbosity_level
 
 #include <chrono>                                   // std::chrono::{steady_clock, duration}
 #include <cstdlib>                                  // EXIT_SUCCESS, EXIT_FAILURE
