+++ src/plssvm/backends/OpenMP/csvm.cpp
@@ -15,7 +15,7 @@
 #include "plssvm/constants.hpp"                                           // plssvm::real_type
 #include "plssvm/csvm.hpp"                                                // plssvm::csvm
 #include "plssvm/detail/assert.hpp"                                       // PLSSVM_ASSERT
-#include "plssvm/detail/logger.hpp"                                       // plssvm::detail::log, plssvm::verbosity_level
+#include "plssvm/detail/logging.hpp"                                      // plssvm::detail::log
 #include "plssvm/detail/memory_size.hpp"                                  // plssvm::detail::memory_size
 #include "plssvm/detail/operators.hpp"                                    // various operator overloads for std::vector and scalars
 #include "plssvm/detail/performance_tracker.hpp"                          // plssvm::detail::tracking_entry, PLSSVM_DETAIL_PERFORMANCE_TRACKER_ADD_TRACKING_ENTRY
@@ -23,6 +23,7 @@
 #include "plssvm/matrix.hpp"                                              // plssvm::aos_matrix, plssvm::soa_matrix
 #include "plssvm/parameter.hpp"                                           // plssvm::parameter
 #include "plssvm/target_platforms.hpp"                                    // plssvm::target_platform
+#include "plssvm/verbosity_levels.hpp"                                    // plssvm::verbosity_level
 
 #include "fmt/chrono.h"   // directly print std::chrono literals with fmt
 #include "fmt/core.h"     // fmt::format
