+++ src/plssvm/backends/SYCL/hipSYCL/csvm.cpp
@@ -19,7 +19,7 @@
 #include "plssvm/backends/SYCL/predict_kernel.hpp"                      // plssvm::sycl::detail::{kernel_w, device_kernel_predict_polynomial, device_kernel_predict_rbf}
 #include "plssvm/constants.hpp"                                         // plssvm::real_type
 #include "plssvm/detail/assert.hpp"                                     // PLSSVM_ASSERT
-#include "plssvm/detail/logger.hpp"                                     // plssvm::detail::log, plssvm::verbosity_level
+#include "plssvm/detail/logging.hpp"                                    // plssvm::detail::log
 #include "plssvm/detail/memory_size.hpp"                                // plssvm::detail::memory_size
 #include "plssvm/detail/performance_tracker.hpp"                        // plssvm::detail::tracking_entry
 #include "plssvm/detail/utility.hpp"                                    // plssvm::detail::get_system_memory
@@ -27,6 +27,7 @@
 #include "plssvm/kernel_function_types.hpp"                             // plssvm::kernel_type
 #include "plssvm/parameter.hpp"                                         // plssvm::parameter, plssvm::detail::parameter
 #include "plssvm/target_platforms.hpp"                                  // plssvm::target_platform
+#include "plssvm/verbosity_levels.hpp"                                  // plssvm::verbosity_level
 
 #include "fmt/color.h"    // fmt::fg, fmt::color::orange
 #include "fmt/core.h"     // fmt::format
