+++ tests/verbosity_levels.cpp
@@ -5,17 +5,17 @@
  * @license This file is part of the PLSSVM project which is released under the MIT license.
  *          See the LICENSE.md file in the project root for full license information.
  *
- * @brief Tests for the logging function.
+ * @brief Tests for the verbosity level enumeration.
  */
 
-#include "plssvm/detail/logger.hpp"
+#include "plssvm/verbosity_levels.hpp"
 
 #include "plssvm/detail/utility.hpp"  // plssvm::detail::to_underlying
 
 #include "custom_test_macros.hpp"  // EXPECT_CONVERSION_TO_STRING, EXPECT_CONVERSION_FROM_STRING
 #include "utility.hpp"             // util::redirect_output
 
-#include "gtest/gtest.h"  // TEST_F, EXPECT_EQ, EXPECT_TRUE, ::testing::Test
+#include "gtest/gtest.h"  // TEST, EXPECT_EQ, EXPECT_TRUE
 
 #include <sstream>  // std::istringstream
 
@@ -101,60 +101,4 @@ TEST(VerbosityLevel, compound_bitwise_and) {
     plssvm::verbosity_level verb = plssvm::verbosity_level::full | plssvm::verbosity_level::libsvm;
     verb &= plssvm::verbosity_level::full;
     EXPECT_EQ(plssvm::detail::to_underlying(verb), 0b100);
-}
-
-class Logger : public ::testing::Test, public util::redirect_output<> {};
-
-TEST_F(Logger, enabled_logging) {
-    // explicitly enable logging
-    plssvm::verbosity = plssvm::verbosity_level::full;
-
-    // log a message
-    plssvm::detail::log(plssvm::verbosity_level::full, "Hello, World!");
-
-    // check captured output
-    EXPECT_EQ(this->get_capture(), "Hello, World!");
-}
-TEST_F(Logger, enabled_logging_with_args) {
-    // explicitly enable logging
-    plssvm::verbosity = plssvm::verbosity_level::full;
-
-    // log a message
-    plssvm::detail::log(plssvm::verbosity_level::full, "int: {}, float: {}, str: {}", 42, 1.5, "abc");
-
-    // check captured output
-    EXPECT_EQ(this->get_capture(), "int: 42, float: 1.5, str: abc");
-}
-
-TEST_F(Logger, disabled_logging) {
-    // explicitly disable logging
-    plssvm::verbosity = plssvm::verbosity_level::quiet;
-
-    // log message
-    plssvm::detail::log(plssvm::verbosity_level::full, "Hello, World!");
-
-    // since logging has been disabled, nothing should have been captured
-    EXPECT_TRUE(this->get_capture().empty());
-}
-TEST_F(Logger, disabled_logging_with_args) {
-    // explicitly disable logging
-    plssvm::verbosity = plssvm::verbosity_level::quiet;
-
-    // log message
-    plssvm::detail::log(plssvm::verbosity_level::full, "int: {}, float: {}, str: {}", 42, 1.5, "abc");
-
-    // since logging has been disabled, nothing should have been captured
-    EXPECT_TRUE(this->get_capture().empty());
-}
-
-TEST_F(Logger, mismatching_verbosity_level) {
-    // set verbosity_level to libsvm
-    plssvm::verbosity = plssvm::verbosity_level::libsvm;
-
-    // log message with full
-    plssvm::detail::log(plssvm::verbosity_level::full, "Hello, World!");
-    plssvm::detail::log(plssvm::verbosity_level::full, "int: {}, float: {}, str: {}", 42, 1.5, "abc");
-
-    // there should not be any output
-    EXPECT_TRUE(this->get_capture().empty());
 }
